chore: small cleanup
