+++ src/App.tsx
@@ -1,138 +1,17 @@
 import React from 'react';
-import { Navigate, Route, Routes } from 'react-router-dom';
+import { Route, Routes } from 'react-router-dom';
 
-import { Form, FormFirstPage } from 'src/components/form/Form';
-import { PresentationComponent } from 'src/components/presentation/Presentation';
 import { ProcessWrapperWrapper } from 'src/components/wrappers/ProcessWrapper';
-import { useApplicationMetadata } from 'src/features/applicationMetadata/ApplicationMetadataProvider';
-import { LayoutValidationProvider } from 'src/features/devtools/layoutValidation/useLayoutValidation';
-import { FormProvider } from 'src/features/form/FormContext';
+import { Entrypoint } from 'src/features/entrypoint/Entrypoint';
 import { InstanceProvider } from 'src/features/instance/InstanceContext';
-import { InstantiateContainer } from 'src/features/instantiate/containers/InstantiateContainer';
-import { NoValidPartiesError } from 'src/features/instantiate/containers/NoValidPartiesError';
 import { PartySelection } from 'src/features/instantiate/containers/PartySelection';
-import { UnknownError } from 'src/features/instantiate/containers/UnknownError';
 import { InstanceSelectionWrapper } from 'src/features/instantiate/selection/InstanceSelection';
-import { useCurrentParty, useCurrentPartyIsValid, useValidParties } from 'src/features/party/PartiesProvider';
-import { useProfile } from 'src/features/profile/ProfileProvider';
-import { useAllowAnonymousIs } from 'src/features/stateless/getAllowAnonymous';
-import { PresentationType } from 'src/types';
-import { useIsStatelessApp } from 'src/utils/useIsStatelessApp';
-import type { ShowTypes } from 'src/features/applicationMetadata';
-
-export const RenderStateless = () => (
-  <FormProvider>
-    <LayoutValidationProvider>
-      <Routes>
-        <Route
-          path=':pageKey'
-          element={
-            <PresentationComponent type={PresentationType.Stateless}>
-              <Form />
-            </PresentationComponent>
-          }
-        />
-        <Route
-          path='*'
-          element={<FormFirstPage />}
-        />
-      </Routes>
-    </LayoutValidationProvider>
-  </FormProvider>
-);
-
-const ShowOrInstantiate: React.FC<{ show: ShowTypes }> = ({ show }) => {
-  if (show === 'select-instance') {
-    return (
-      <Navigate
-        to={'/instance-selection'}
-        replace={true}
-      />
-    );
-  }
-
-  if (show === 'new-instance') {
-    return <InstantiateContainer />;
-  }
-
-  if (show === 'stateless') {
-    return <RenderStateless />;
-  }
-
-  if (show === 'stateless-anon') {
-    return <RenderStateless />;
-  }
-
-  window.logErrorOnce('Unknown applicationMetadata.onEntry type:', show);
-
-  return <UnknownError />;
-};
-
-export const DefaultComponent = () => {
-  const applicationMetadata = useApplicationMetadata();
-  const show: ShowTypes = applicationMetadata.onEntry?.show ?? 'new-instance';
-  const validParties = useValidParties();
-  const profile = useProfile();
-  const partyIsValid = useCurrentPartyIsValid();
-  const isStateless = useIsStatelessApp();
-  const party = useCurrentParty();
-  const allowAnonymous = useAllowAnonymousIs(true);
-
-  if (isStateless && allowAnonymous) {
-    if (party === undefined) {
-      return <RenderStateless />;
-    }
-  }
-
-  if (!partyIsValid) {
-    return (
-      <Navigate
-        to={'/party-selection/403'}
-        replace={true}
-      />
-    );
-  }
-
-  if (!validParties?.length) {
-    return <NoValidPartiesError />;
-  }
-
-  if (validParties?.length === 1) {
-    return <ShowOrInstantiate show={show} />;
-  }
-
-  if (validParties?.length && validParties?.length > 1) {
-    if (applicationMetadata.promptForParty === 'always') {
-      return (
-        <Navigate
-          to={'/party-selection/explained'}
-          replace={true}
-        />
-      );
-    }
-
-    if (applicationMetadata.promptForParty === 'never') {
-      return <ShowOrInstantiate show={show} />;
-    }
-
-    if (profile?.profileSettingPreference.doNotPromptForParty) {
-      return <ShowOrInstantiate show={show} />;
-    }
-    return (
-      <Navigate
-        to={'/party-selection/explained'}
-        replace={true}
-      />
-    );
-  }
-  return <UnknownError />;
-};
 
 export const App = () => (
   <Routes>
     <Route
       path={'*'}
-      element={<DefaultComponent />}
+      element={<Entrypoint />}
     />
     <Route
       path='/instance-selection/*'
