+++ src/features/entrypoint/Entrypoint.tsx
@@ -1,73 +1,125 @@
 import React from 'react';
-import { Navigate, Outlet, Route, Routes, useLocation } from 'react-router-dom';
+import { Navigate, Route, Routes } from 'react-router-dom';
 
 import { Form, FormFirstPage } from 'src/components/form/Form';
 import { PresentationComponent } from 'src/components/presentation/Presentation';
+import { useApplicationMetadata } from 'src/features/applicationMetadata/ApplicationMetadataProvider';
 import { LayoutValidationProvider } from 'src/features/devtools/layoutValidation/useLayoutValidation';
 import { FormProvider } from 'src/features/form/FormContext';
-import { useCurrentParty, useCurrentPartyIsValid } from 'src/features/party/PartiesProvider';
+import { InstantiateContainer } from 'src/features/instantiate/containers/InstantiateContainer';
+import { NoValidPartiesError } from 'src/features/instantiate/containers/NoValidPartiesError';
+import { UnknownError } from 'src/features/instantiate/containers/UnknownError';
+import { useCurrentParty, useCurrentPartyIsValid, useValidParties } from 'src/features/party/PartiesProvider';
+import { useProfile } from 'src/features/profile/ProfileProvider';
 import { useAllowAnonymousIs } from 'src/features/stateless/getAllowAnonymous';
-import { usePromptForParty } from 'src/hooks/usePromptForParty';
 import { PresentationType } from 'src/types';
 import { useIsStatelessApp } from 'src/utils/useIsStatelessApp';
+import type { ShowTypes } from 'src/features/applicationMetadata';
 
-export function Entrypoint() {
+const RenderStateless = () => (
+  <FormProvider>
+    <LayoutValidationProvider>
+      <Routes>
+        <Route
+          path=':pageKey'
+          element={
+            <PresentationComponent type={PresentationType.Stateless}>
+              <Form />
+            </PresentationComponent>
+          }
+        />
+        <Route
+          path='*'
+          element={<FormFirstPage />}
+        />
+      </Routes>
+    </LayoutValidationProvider>
+  </FormProvider>
+);
+
+const ShowOrInstantiate: React.FC<{ show: ShowTypes }> = ({ show }) => {
+  if (show === 'select-instance') {
+    return (
+      <Navigate
+        to={'/instance-selection'}
+        replace={true}
+      />
+    );
+  }
+
+  if (show === 'new-instance') {
+    return <InstantiateContainer />;
+  }
+
+  if (show === 'stateless') {
+    return <RenderStateless />;
+  }
+
+  if (show === 'stateless-anon') {
+    return <RenderStateless />;
+  }
+
+  window.logErrorOnce('Unknown applicationMetadata.onEntry type:', show);
+
+  return <UnknownError />;
+};
+
+export const Entrypoint = () => {
+  const applicationMetadata = useApplicationMetadata();
+  const show: ShowTypes = applicationMetadata.onEntry?.show ?? 'new-instance';
+  const validParties = useValidParties();
+  const profile = useProfile();
+  const partyIsValid = useCurrentPartyIsValid();
   const isStateless = useIsStatelessApp();
   const party = useCurrentParty();
-  const partyIsValid = useCurrentPartyIsValid();
   const allowAnonymous = useAllowAnonymousIs(true);
-  const alwaysPromptForParty = usePromptForParty();
 
-  const location = useLocation();
-  const isMissingParty = party === undefined && !allowAnonymous;
+  if (isStateless && allowAnonymous) {
+    if (party === undefined) {
+      return <RenderStateless />;
+    }
+  }
 
-  if (alwaysPromptForParty) {
+  if (!partyIsValid) {
     return (
       <Navigate
-        to={`/party-selection`}
+        to={'/party-selection/403'}
         replace={true}
       />
     );
   }
 
-  if (!partyIsValid || isMissingParty) {
-    if (location.pathname.includes('party-selection')) {
-      return <Outlet />;
+  if (!validParties?.length) {
+    return <NoValidPartiesError />;
+  }
+
+  if (validParties?.length === 1) {
+    return <ShowOrInstantiate show={show} />;
+  }
+
+  if (validParties?.length && validParties?.length > 1) {
+    if (applicationMetadata.promptForParty === 'always') {
+      return (
+        <Navigate
+          to={'/party-selection/explained'}
+          replace={true}
+        />
+      );
     }
 
+    if (applicationMetadata.promptForParty === 'never') {
+      return <ShowOrInstantiate show={show} />;
+    }
+
+    if (profile?.profileSettingPreference.doNotPromptForParty) {
+      return <ShowOrInstantiate show={show} />;
+    }
     return (
       <Navigate
-        to={`/party-selection/403`}
+        to={'/party-selection/explained'}
         replace={true}
       />
     );
   }
-
-  // Stateless view
-  if (isStateless) {
-    return (
-      <FormProvider>
-        <LayoutValidationProvider>
-          <Routes>
-            <Route
-              path=':pageKey'
-              element={
-                <PresentationComponent type={PresentationType.Stateless}>
-                  <Form />
-                </PresentationComponent>
-              }
-            />
-            <Route
-              path='*'
-              element={<FormFirstPage />}
-            />
-          </Routes>
-        </LayoutValidationProvider>
-      </FormProvider>
-    );
-  }
-
-  console.log('outlet');
-
-  return <Outlet />;
-}
+  return <UnknownError />;
+};
